client/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
asyncio==3.4.3
attrs==25.3.0
blinker==1.9.0
browserforge==1.2.3
camoufox==0.4.11
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
frozenlist==1.5.0
geoip2==5.0.1
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
language-tags==1.2.0
lxml==5.3.2
MarkupSafe==3.0.2
maxminddb==2.6.3
multidict==6.2.0
numpy==2.2.4
orjson==3.10.16
patchright==1.51.3
platformdirs==4.3.7
playwright==1.51.0
priority==2.0.0
propcache==0.3.1
psutil==7.0.0
pyee==12.1.1
PySocks==1.7.1
PyYAML==6.0.2
Quart==0.20.0
requests==2.32.3
screeninfo==0.8.1
tqdm==4.67.1
typing_extensions==4.13.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.19.0